+++ examples/sync_demo.py
@@ -8,7 +8,7 @@ def sync_demo():
         "engine": "candle",
         "kind": "stable-lm",
         "model_id": "lmz/candle-stablelm-3b-4e1t",
-        "sample_len": 1024,
+        "sample_len": 512,
         "prompt": "Building a website can be done in 10 simple steps:\nStep 1:"
     }))
 
@@ -16,7 +16,7 @@ def sync_demo():
         "engine": "candle",
         "kind": "stable-lm",
         "model_id": "lmz/candle-stablelm-3b-4e1t",
-        "sample_len": 1024,
+        "sample_len": 512,
         "messages": [
             {"role": "system", "content": "You are a helpful assistant."},
             {"role": "user", "content": "I need help building a website."},
@@ -29,16 +29,16 @@ def sync_demo():
         "engine": "candle",
         "kind": "stable-lm",
         "model_id": "lmz/candle-stablelm-3b-4e1t",
-        "sample_len": 1024,
+        "sample_len": 512,
         "prompt": "Building a website can be done in 10 simple steps:\nStep 1:"
     }):
-        print(chunk)
+        print(chunk, sep='', end='', flush=True)
 
     for chunk in sync_client.iter_chat(**{
         "engine": "candle",
         "kind": "stable-lm",
         "model_id": "lmz/candle-stablelm-3b-4e1t",
-        "sample_len": 1024,
+        "sample_len": 512,
         "messages": [
             {"role": "system", "content": "You are a helpful assistant."},
             {"role": "user", "content": "I need help building a website."},
@@ -46,7 +46,7 @@ def sync_demo():
             {"role": "user", "content": "Building a website can be done in 10 simple steps. Explain step by step."}
         ]
     }):
-        print(chunk)
+        print(chunk, sep='', end='', flush=True)
 
 
 if __name__ == '__main__':